+++ NEWS.md
@@ -4,9 +4,11 @@ Updated to Julia v0.7 (older versions not supported)
 
 ##### Changes
 * update to Julia v0.7+ and drop support for older versions [#42], [#44]
+* improve FileIO integration [#46]
 
 [#42]: https://github.com/JuliaStats/RData.jl/issues/42
 [#44]: https://github.com/JuliaStats/RData.jl/issues/44
+[#46]: https://github.com/JuliaStats/RData.jl/issues/46
 
 ## RData v0.4.0 Release Notes